Make Method luckyDrawMessages(List<AbstractMessage>) abstract and provide default implementation in DefaultHighPriest; this would allow multiple implementation to be provided in case luckyDrawMessages behavior has to change
